user register, login, logout added

+++ src/main/java/pl/akademiakodu/blogApplication/ropository/PostRepository.java
@@ -4,7 +4,12 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 import pl.akademiakodu.blogApplication.model.entities.Post;
 
+import java.util.ArrayList;
+import java.util.List;
+
 @Repository
-public interface PostRepository extends JpaRepository<Post, Integer> {
+public interface PostRepository extends JpaRepository<Post, Long> {
+
+
 
 }
